requirements.txt
This package requires Python 3.8+
torch
numpy
pyyaml
pillow
transformers
